requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
DateTime==5.5
Django==5.0.4
feedparser==6.0.11
pillow==10.3.0
pytz==2024.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
zope.interface==6.2
